+++ lib/run_loop/core_simulator.rb
@@ -46,6 +46,9 @@ class RunLoop::CoreSimulator
   attr_reader :xcrun
 
   # @!visibility private
+  attr_reader :sim_keyboard
+
+  # @!visibility private
   METADATA_PLIST = '.com.apple.mobile_container_manager.metadata.plist'
 
   # @!visibility private
@@ -401,6 +404,11 @@ class RunLoop::CoreSimulator
   end
 
   # @!visibility private
+  def sim_keyboard
+    @sim_keyboard ||= RunLoop::SimKeyboardSettings.new(device)
+  end
+
+  # @!visibility private
   def simctl
     @simctl ||= RunLoop::Simctl.new
   end
@@ -424,7 +432,7 @@ class RunLoop::CoreSimulator
 
     RunLoop::CoreSimulator.quit_simulator
     RunLoop::CoreSimulator.ensure_hardware_keyboard_connected(pbuddy)
-    device.simulator_ensure_software_keyboard_will_show
+    sim_keyboard.ensure_soft_keyboard_will_show
 
     args = ['open', '-g', '-a', sim_app_path, '--args',
             '-CurrentDeviceUDID', device.udid,
@@ -445,11 +453,6 @@ class RunLoop::CoreSimulator
 
     options = { :timeout => 5, :raise_on_timeout => true }
     RunLoop::ProcessWaiter.new(sim_name, options).wait_for_any
-
-    # open -g no longer launches application in the background. We want the
-    # Simulator to open in the background because when it is opened in the
-    # foreground, it steals (key application) focus which is disruptive.
-    send_simulator_to_background
 
     if merged_options[:wait_for_stable]
       device.simulator_wait_for_stable_state
@@ -679,23 +682,6 @@ Command had no output.
   end
 
   # @!visibility private
-  def send_simulator_to_background
-    script = "tell application \"System Events\" to tell process \"#{sim_name}\" to set visible to false"
-    begin
-      system("osascript",  "-e", script)
-    rescue => _
-      RunLoop.log_debug("Could not put simulator into the background")
-    end
-
-    script = "tell application \"System Events\" to tell process \"#{sim_name}\" to set visible to true"
-    begin
-      system("osascript",  "-e", script)
-    rescue => _
-      RunLoop.log_debug("Could not put simulator into the foreground")
-    end
-  end
-
-  # @!visibility private
   def uninstall_app_with_simctl
     launch_simulator
 
@@ -824,7 +810,7 @@ Command had no output.
       return true
     end
 
-    if !device.simulator_software_keyboard_will_show?
+    if !sim_keyboard.soft_keyboard_will_show?
       RunLoop.log_debug("Simulator relaunch required:  software keyboard is minimized")
       return true
     end